Renamed websocket modules. Documented problems with main.py. Game side not integrated yet.
